requirements.txt
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
django-filter==2.2.0
django-formtools==2.2
django-guardian==2.2.0
django-import-export==2.0.2
django-mdeditor==0.1.17
django-ranged-response==0.2.0
django-reversion==3.0.7
django-simple-captcha==0.5.12
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.17.0
idna==2.9
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
Markdown==3.2.1
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==1.4.6
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.0.0
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.2
sqlparse==0.3.1
tablib==1.1.0
uritemplate==3.0.1
urllib3==1.25.8
xlrd==1.2.0
XlsxWriter==1.2.8
xlwt==1.3.0
